requirements.txt
librosa==0.9.2
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.1
Pillow==9.4.0
scipy==1.10.0
seaborn==0.12.2
streamlit==1.16.0
tensorflow==2.11.0
